Updated READ.ME to reflect aws UI/behavior changes
The instructions for uploading the built website code to the s3 bucket is out of date.  See: http://docs.aws.amazon.com/AmazonS3/latest/UG/UploadingObjectsintoAmazonS3.html

These instructions are now updated to work with the new UI.  Currently if you select 'Upload' then try to upload all the contents of website/dist it will only pick up on files, not directories.  The only way to upload a whole folder with files and directories that I could find is by drag and dropping the content in.